requirements.txt
pandas
python-dotenv
torch
transformers
langchain
langchain-chroma
langchain-huggingface
langchain-openai
openai
pandas